# Add BufWriter to ChildStdin on the plugin interface