# Description
This speeds up writing messages to the plugin, because otherwise every individual piece of the messages (not even the entire message) is written with one syscall, leading to a lot of back and forth with the kernel.

I learned this by running `strace` to debug something and saw a ton of `write()` calls.

```nushell
# Before
1..10 | each { timeit { example seq 1 10000 | example sum } } | math avg
269ms 779µs 149ns
# After
> 1..10 | each { timeit { example seq 1 10000 | example sum } } | math avg
39ms 636µs 643ns
```

# User-Facing Changes
- Performance improvement

# Tests + Formatting
- :green_circle: `toolkit fmt`
- :green_circle: `toolkit clippy`
- :green_circle: `toolkit test`
- :green_circle: `toolkit test stdlib`