@WindSoilder 2024-04-06
What if input is not enough to OUTPUT_BUFFER_SIZE?  Will it hang on forever?