@devyn 2024-04-06
> What if input is not enough to OUTPUT_BUFFER_SIZE? Will it hang on forever?

No, there are enough `flush()` calls already where it matters to prevent that. Plus, `Stdout` is already buffered, so this is just for `ChildStdin`.